+++ packages/components/src/PieChart/PieChart.scss
@@ -16,7 +16,7 @@ $tc-pie-chart-colors: (rio-grande $rio-grande, chestnut-rose $chestnut-rose, lig
 }
 
 .tc-pie-chart-loading {
-	display: flex;
+	display: inline-flex;
 	align-items: center;
 	min-height: $tc-pie-chart-loading-min-height;
 
@@ -62,8 +62,7 @@ $tc-pie-chart-colors: (rio-grande $rio-grande, chestnut-rose $chestnut-rose, lig
 	}
 }
 
-.tc-pie-chart,
-.tc-pie-chart-loading {
+.tc-pie-chart {
 	font-size: $tc-pie-chart-font-size;
 	display: flex;
 	align-items: center;
