+++ packages/components/src/PieChart/index.js
@@ -1,4 +1,4 @@
 import PieChartButton from './PieChartButton.component';
 import PieChart from './PieChart.component';
 
-export { PieChartButton as default, PieChart };
+export { PieChartButton, PieChart };
